+++ app/controllers/notes_controller.rb
@@ -1,25 +1,22 @@
 class NotesController < ApplicationController
   before_action :find_project
-  before_action :find_notes, only: :index
-  after_action :verify_authorized
-
-  def index
-    authorize Note
-  end
+  before_action :find_project_level_notes, only: :index
+  after_action :verify_authorized, except: :index
+  after_action :verify_policy_scoped, only: :index
 
   def new
-    authorize Note
-    @note = Note.new(project: @project, user_id:)
+    @note = Note.new(project: @project, user_id:, task_id: params[:task_id])
+    authorize @note
   end
 
   def create
-    authorize Note
     @note = Note.new(project: @project, user_id:, **note_params)
+    authorize @note
 
     if @note.valid?
       @note.save
 
-      redirect_to project_notes_path(@project), notice: I18n.t("note.create.success")
+      redirect_to(redirect_path, notice: I18n.t("note.create.success"))
     else
       render :new
     end
@@ -38,7 +35,7 @@ class NotesController < ApplicationController
 
     if @note.valid?
       @note.save
-      redirect_to project_notes_path(@project), notice: I18n.t("note.update.success")
+      redirect_to redirect_path, notice: I18n.t("note.update.success")
     else
       render :edit
     end
@@ -50,7 +47,7 @@ class NotesController < ApplicationController
 
     @note.destroy
 
-    redirect_to project_notes_path(@project), notice: I18n.t("note.destroy.success")
+    redirect_to redirect_path, notice: I18n.t("note.destroy.success")
   end
 
   def confirm_destroy
@@ -58,15 +55,19 @@ class NotesController < ApplicationController
     authorize @note
   end
 
+  private def redirect_path
+    @note.task_level_note? ? project_task_path(@project, @note.task.id) : project_notes_path(@project)
+  end
+
   private def find_project
     @project = Project.find(params[:project_id])
   end
 
-  private def find_notes
-    @notes = Note.includes([:user]).where(project: @project)
+  private def find_project_level_notes
+    @notes = policy_scope(Note).includes([:user]).project_level_notes(@project)
   end
 
   private def note_params
-    params.require(:note).permit(:body)
+    params.require(:note).permit(:body, :task_id)
   end
 end