+++ app/policies/note_policy.rb
@@ -6,10 +6,6 @@ class NotePolicy
     @record = note
   end
 
-  def index?
-    true
-  end
-
   def create?
     true
   end
@@ -32,5 +28,20 @@ class NotePolicy
 
   def confirm_destroy?
     edit?
+  end
+
+  class Scope
+    def initialize(user, scope)
+      @user = user
+      @scope = scope
+    end
+
+    def resolve
+      scope.all
+    end
+
+    private
+
+    attr_reader :user, :scope
   end
 end