+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2022_09_28_143711) do
+ActiveRecord::Schema[7.0].define(version: 2022_09_29_115655) do
   create_table "actions", id: :uuid, default: -> { "newid()" }, force: :cascade do |t|
     t.string "title", null: false
     t.integer "order", null: false
@@ -44,7 +44,9 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_28_143711) do
     t.uuid "user_id"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
+    t.uuid "task_id"
     t.index ["project_id"], name: "index_notes_on_project_id"
+    t.index ["task_id"], name: "index_notes_on_task_id"
     t.index ["user_id"], name: "index_notes_on_user_id"
   end
 
@@ -107,6 +109,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_28_143711) do
   add_foreign_key "actions", "tasks"
   add_foreign_key "contacts", "projects"
   add_foreign_key "notes", "projects"
+  add_foreign_key "notes", "tasks"
   add_foreign_key "notes", "users"
   add_foreign_key "projects", "users", column: "caseworker_id"
   add_foreign_key "projects", "users", column: "regional_delivery_officer_id"