+++ spec/models/note_spec.rb
@@ -8,6 +8,7 @@ RSpec.describe Note, type: :model do
   describe "Relationships" do
     it { is_expected.to belong_to(:project) }
     it { is_expected.to belong_to(:user) }
+    it { is_expected.to belong_to(:task).required(false) }
   end
 
   describe "Validations" do
@@ -18,10 +19,10 @@ RSpec.describe Note, type: :model do
   end
 
   describe "Scopes" do
+    before { mock_successful_api_responses(urn: any_args, ukprn: any_args) }
+
     describe "default_scope" do
       before do
-        mock_successful_api_responses(urn: any_args, ukprn: any_args)
-
         freeze_time
         travel_to(Date.yesterday) { create(:note, body: "Yesterday's note.") }
         create(:note, body: "Today's note.")
@@ -30,6 +31,36 @@ RSpec.describe Note, type: :model do
       it "orders descending by the 'created_at' attribute" do
         expect(Note.first.body).to eq "Today's note."
       end
+    end
+
+    describe "project_level_notes" do
+      let!(:project_level_note) { create(:note) }
+      let!(:task_level_note) { create(:note, :task_level_note) }
+
+      subject { Note.project_level_notes(project_level_note.project) }
+
+      it "returns only project level notes" do
+        expect(subject).to include project_level_note
+        expect(subject).not_to include task_level_note
+      end
+    end
+  end
+
+  describe "#task_level_note?" do
+    before { mock_successful_api_responses(urn: any_args, ukprn: any_args) }
+
+    subject { note.task_level_note? }
+
+    context "when the Note is not associated with a Task" do
+      let(:note) { create(:note) }
+
+      it { expect(subject).to be false }
+    end
+
+    context "when the Note is associated with a Task" do
+      let(:note) { create(:note, :task_level_note) }
+
+      it { expect(subject).to be true }
     end
   end
 end