+++ spec/models/task_spec.rb
@@ -12,6 +12,7 @@ RSpec.describe Task, type: :model do
   describe "Relationships" do
     it { is_expected.to belong_to(:section) }
     it { is_expected.to have_many(:actions).dependent(:destroy) }
+    it { is_expected.to have_many(:notes).dependent(:destroy) }
   end
 
   describe "Scopes" do