+++ spec/requests/notes_controller_spec.rb
@@ -12,6 +12,8 @@ RSpec.describe NotesController, type: :request do
   describe "#index" do
     let(:project) { create(:project) }
     let(:project_id) { project.id }
+    let!(:project_level_note) { create(:note, project: project) }
+    let!(:task_level_note) { create(:note, :task_level_note, project: project) }
 
     subject(:perform_request) do
       get project_notes_path(project_id)
@@ -26,6 +28,8 @@ RSpec.describe NotesController, type: :request do
 
     it "returns a successful response" do
       expect(subject).to have_http_status :success
+      expect(response.body).to include project_level_note.body
+      expect(response.body).not_to include task_level_note.body
     end
   end
 
@@ -52,11 +56,12 @@ RSpec.describe NotesController, type: :request do
   describe "#create" do
     let(:project) { create(:project) }
     let(:project_id) { project.id }
-    let(:mock_note) { build(:note) }
+    let(:mock_note) { build(:note, project: project) }
     let(:new_note_body) { "Just had an interesting chat about building regulations." }
+    let(:params) { {note: {body: new_note_body}} }
 
     subject(:perform_request) do
-      post project_notes_path(project_id), params: {note: {body: new_note_body}}
+      post project_notes_path(project_id), params: params
       response
     end
 
@@ -84,6 +89,21 @@ RSpec.describe NotesController, type: :request do
 
         expect(Note.count).to be 1
         expect(Note.last.body).to eq(new_note_body)
+        expect(Note.last.task).to be_nil
+      end
+
+      context "when the note is a task level note" do
+        let(:task) { create(:task) }
+        let(:params) { {note: {body: new_note_body, task_id: task.id}} }
+
+        it "saves the note and redirects to the task view with a success message" do
+          expect(subject).to redirect_to(project_task_path(project.id, task.id))
+          expect(request.flash[:notice]).to eq(I18n.t("note.create.success"))
+
+          expect(Note.count).to be 1
+          expect(Note.last.body).to eq(new_note_body)
+          expect(Note.last.task).to eq task
+        end
       end
     end
   end
@@ -134,9 +154,10 @@ RSpec.describe NotesController, type: :request do
     let(:note) { create(:note, user: user) }
     let(:note_id) { note.id }
     let(:new_note_body) { "This is an updated note body" }
+    let(:params) { {note: {body: new_note_body}} }
 
     subject(:perform_request) do
-      put project_note_path(project_id, note_id), params: {note: {body: new_note_body}}
+      put project_note_path(project_id, note_id), params: params
       response
     end
 
@@ -183,6 +204,20 @@ RSpec.describe NotesController, type: :request do
         expect(Note.count).to be 1
         expect(Note.last.body).to eq(new_note_body)
       end
+
+      context "when the note is a task level note" do
+        let(:task) { create(:task) }
+        let(:params) { {note: {body: new_note_body, task_id: task.id}} }
+
+        it "saves the note and redirects to the task view with a success message" do
+          expect(subject).to redirect_to(project_task_path(project.id, task.id))
+          expect(request.flash[:notice]).to eq(I18n.t("note.update.success"))
+
+          expect(Note.count).to be 1
+          expect(Note.last.body).to eq(new_note_body)
+          expect(Note.last.task).to eq task
+        end
+      end
     end
   end
 
@@ -214,6 +249,17 @@ RSpec.describe NotesController, type: :request do
       expect(request.flash[:notice]).to eq(I18n.t("note.destroy.success"))
 
       expect(Note.where(id: note_id)).to_not exist
+    end
+
+    context "when the note is a task level note" do
+      let(:note) { create(:note, :task_level_note, user: user) }
+
+      it "deletes the note and redirects to the task view with a success message" do
+        expect(subject).to redirect_to(project_task_path(project.id, note.task.id))
+        expect(request.flash[:notice]).to eq(I18n.t("note.destroy.success"))
+
+        expect(Note.where(id: note_id)).to_not exist
+      end
     end
   end