+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Users can add, edit, and delete task level notes. These are displayed on the
+  task page.
+
+### Changed
+
+- Only project level notes are shown on the notes page.
+- The back link destination is the task page for task level notes, or the notes
+  page for project level notes.
+
 ## [Release 5][release-5]
 
 ### Changed